requirements.txt
numpy==1.24.3
opencv-python==4.8.0.76
pillow==10.0.0
scikit-image==0.21.0
matplotlib==3.7.2
torch==2.0.1
torchvision==0.15.2
# Using Tkinter for UI (built-in with Python)
